app/analytics/requirements.txt
torch>=2.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
sqlalchemy>=2.0.0
pandas>=2.0.0
numpy>=1.23.0
scikit-learn>=1.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.5
